+++ glaze-lib/main.rkt
@@ -5,9 +5,11 @@
          "browser.rkt"
          "assets.rkt"
          "build.rkt"
+         "app.rkt"
          "tray/main.rkt"
          "webview/main.rkt")
 
 (provide (all-from-out "server.rkt" "api.rkt" "browser.rkt" "assets.rkt" "build.rkt")
+         (all-from-out "app.rkt")
          (all-from-out "tray/main.rkt")
          (all-from-out "webview/main.rkt"))